origin

+++ controllers/booking/booking.js
@@ -5,6 +5,8 @@ import roomModel from '../../models/rooms.js';
 import incomeModel from '../../models/income.js';
 import taxModel from '../../models/tax.model.js';
 import mealModel from '../../models/meal.js';
+import { ledgerModel } from "../../models/ledger.model.js"
+import { ledgerHistoryModel } from "../../models/ledgerHistory.model.js"
 
 import Joi from "joi";
 
@@ -13,55 +15,59 @@ const bookingValidationSchema = Joi.object({
     firstName: Joi.string().required(),
     lastName: Joi.string().required(),
     gender: Joi.string().valid("male", "female", "other").required(),
-    meal: Joi.string().required(),
+    meal: Joi.string().optional().allow(""),
     phone: Joi.string().required(),
     email: Joi.string().email().required(),
     address: Joi.string().required(),
     roomType: Joi.string().required(),
-    packages: Joi.string().optional().allow(""), 
+    packages: Joi.string().optional().allow(""),
     roomNo: Joi.string().required(),
     arrivedDate: Joi.date().optional(),
     departDate: Joi.date().required(),
     totalPerson: Joi.number().integer().min(1).required(),
-    tax: Joi.string().required(),
-    note: Joi.string().optional(),
+    tax: Joi.string().optional().allow(""),
+    note: Joi.string().allow(""),
+    bookingSourceId: Joi.string().allow(""),
 });
 
 export const addBooking = async (req, res) => {
-    const { 
-        firstName, 
-        lastName, 
-        gender, 
-        phone, 
-        meal, 
-        email, 
-        address, 
-        roomType, 
-        packages, 
-        roomNo, 
-        arrivedDate, 
-        departDate, 
-        totalPerson, 
-        note, 
-        tax 
+    const {
+        firstName,
+        lastName,
+        gender,
+        phone,
+        meal,
+        email,
+        address,
+        roomType,
+        packages,
+        roomNo,
+        arrivedDate,
+        departDate,
+        totalPerson,
+        note,
+        tax,
+        bookingSourceId
     } = req.body;
 
-
     const adminId = req.user.adminId || req.user._id;
-    const documentThumbnail =req.files?.documentThumbnail?.map(file => file.filename) || [];
-    console.log(documentThumbnail)
+    const documentThumbnail = req.files?.documentThumbnail?.map(file => file.filename) || [];
+
     const { error } = bookingValidationSchema.validate(req.body);
     if (error) {
-        return res.status(400).json({ message: "Validation error", errors: error.details });
+        return res.status(400).json({ success: false, message: "Validation error", errors: error.details });
     }
 
     try {
         // Check if room type and room number are valid
         const selectedRoomType = await roomCategoryModel.findById(roomType);
         const selectedRoom = await roomModel.findById(roomNo);
-
+        const ledger = await ledgerModel.findOne({ closingBalance: null });
+        if (!ledger) {
+            return res.status(401).json({ success: false, message: "Open Ledger First" });
+        }
         if (!selectedRoom || selectedRoom.status !== "open") {
-            return res.status(400).json({ message: "Room is not available" });
+            return res.status(400).json({ success: false, message: "Room is not available" });
         }
 
         const arrivalDate = new Date(arrivedDate);
@@ -69,7 +75,7 @@ export const addBooking = async (req, res) => {
         const numberOfNights = (departureDate - arrivalDate) / (1000 * 60 * 60 * 24);
 
         if (numberOfNights <= 0) {
-            return res.status(400).json({ message: "Invalid dates: departure date must be after arrival date" });
+            return res.status(400).json({ success: false, message: "Invalid dates: departure date must be after arrival date" });
         }
 
         const roomPricePerNight = selectedRoom.rent;
@@ -84,42 +90,48 @@ export const addBooking = async (req, res) => {
         }
 
         // Retrieve meal price based on selected meal
-        const selectedMeal = await mealModel.findById(meal);
-        const mealPrice = selectedMeal.price ? selectedMeal.price : 0; 
+        let mealPrice = 0;
+        if (meal) {
+            const selectedMeal = await mealModel.findById(meal);
+            mealPrice = selectedMeal?.price || 0;
+        }
 
         // Calculate total booking amount
         let totalBookingAmount = (roomPricePerNight * numberOfNights) + (packagePrice * numberOfNights) + (mealPrice * totalPerson);
 
-        // Retrieve tax details
-        const selectedTax = await taxModel.findById(tax);
-        if (selectedTax) {
-            if (selectedTax.type === "fixed") {
-                totalBookingAmount += selectedTax.amount; 
-            } else if (selectedTax.type === "percentage") {
-                totalBookingAmount += (totalBookingAmount * (selectedTax.amount / 100));
+        // Retrieve tax details if tax is provided
+        if (tax) {
+            const selectedTax = await taxModel.findById(tax);
+            if (selectedTax) {
+                if (selectedTax.type === "fixed") {
+                    totalBookingAmount += selectedTax.amount;
+                } else if (selectedTax.type === "percentage") {
+                    totalBookingAmount += (totalBookingAmount * (selectedTax.amount / 100));
+                }
             }
         }
 
         // Create the new booking object
         const newBooking = new bookingModel({
-            adminId,
-            firstName,
-            lastName,
-            gender,
-            phone,
-            meal,
-            email,
-            address,
-            documentThumbnail,
+            adminId: adminId,
+            firstName: firstName,
+            lastName: lastName,
+            gender: gender,
+            phone: phone,
+            meal: meal || null,
+            email: email,
+            address: address,
+            documentThumbnail: documentThumbnail,
             packages: packages || null,
-            roomType,
-            roomNo,
-            arrivedDate,
-            departDate,
-            totalPerson,
-            tax,
-            note,
+            roomType: roomType,
+            roomNo: roomNo,
+            arrivedDate: arrivalDate,
+            departDate: departDate,
+            totalPerson: totalPerson,
+            tax: tax || null,
+            note: note,
             bookingAmount: totalBookingAmount,
+            bookingSourceId: bookingSourceId || null
         });
 
         // Save the booking
@@ -138,20 +150,22 @@ export const addBooking = async (req, res) => {
         });
         await newIncome.save();
 
+        
+        await ledgerHistoryModel.create({
+            ledgerId: ledger._id,
+            type: "Booking",
+            credit: totalBookingAmount,
+            description: `Income from booking: ${firstName} ${lastName} in room number ${roomNo} `,
+        })
         // Update package if applicable
-        if (packages) {
-            await packageModel.updateOne(
-                { _id: packages },
-                { $inc: { numberOfPackage: 1 } }
-            );
-        }
+        
 
-        res.status(201).json({ message: "Booking added successfully", booking: newBooking });
+        res.status(201).json({ success: true, message: "Booking added successfully", booking: newBooking });
     } catch (error) {
-        res.status(500).json({ message: "Failed to add booking", error: error.message });
+        console.log(error)
+        res.status(500).json({ success: false, message: "Internal server error", });
     }
 };
-
 
 
 
@@ -159,8 +173,9 @@ export const addBooking = async (req, res) => {
 const populateBookingDetails = () => [
     { path: 'packages', select: 'packages price' },
     { path: 'roomType', select: 'type' },
-    { path: 'roomNo', select: 'roomNumber rent' },
+    { path: 'roomNo', select: 'roomNumber rent status' },
     { path: 'meal', select: 'name price' },
+    { path: "bookingSourceId", select: "bookingSourceName description" }
 
 ];
 
@@ -181,22 +196,23 @@ export const getBookingsByAdminId = async (req, res) => {
 export const updateBooking = async (req, res) => {
     const { id } = req.params;
 
-    const { 
-        firstName, 
-        lastName, 
-        gender, 
-        phone, 
-        email, 
-        address, 
-        packages, 
-        roomType, 
-        roomNo, 
-        arrivedDate, 
-        departDate, 
-        totalPerson, 
-        note, 
+    const {
+        firstName,
+        lastName,
+        gender,
+        phone,
+        email,
+        address,
+        packages,
+        roomType,
+        roomNo,
+        arrivedDate,
+        departDate,
+        totalPerson,
+        note,
         meal,
-        tax 
+        tax,
+        bookingSourceId
     } = req.body;
     const documentThumbnail = req.files;
 
@@ -214,13 +230,13 @@ export const updateBooking = async (req, res) => {
         // Calculate booking amount
         const roomRentPerNight = room.rent;
         const numberOfNights = (new Date(departDate) - new Date(arrivedDate)) / (1000 * 60 * 60 * 24);
-        
+
         // Calculate total amount
         let totalAmount = (roomRentPerNight * numberOfNights) + (packageDetails.price * numberOfNights);
 
         // Calculate meal price if meal is selected
-        const mealPrice = selectedMeal.price ? selectedMeal.price : 0; 
-        totalAmount += mealPrice * totalPerson; 
+        const mealPrice = selectedMeal.price ? selectedMeal.price : 0;
+        totalAmount += mealPrice * totalPerson;
 
         // Calculate tax amount if tax is selected
         if (selectedTax) {
@@ -247,7 +263,8 @@ export const updateBooking = async (req, res) => {
             departDate,
             totalPerson,
             note,
-            bookingAmount: totalAmount
+            bookingAmount: totalAmount,
+            bookingSourceId: bookingSourceId
         }, {
             new: true,
             runValidators: true
@@ -280,6 +297,15 @@ export const deleteBooking = async (req, res) => {
 
     try {
         const booking = await bookingModel.findById(id)
+        // Find the room associated with the booking
+        const room = await roomModel.findById(booking.roomNo);
+        if (!room) {
+            return res.status(404).json({ success: false, message: "Room not found" });
+        }
+
+        // Update room status to "open"
+        room.status = "open";
+        await room.save();
 
         const deletedBooking = await bookingModel.findOneAndDelete(id)
 
@@ -300,6 +326,42 @@ export const deleteBooking = async (req, res) => {
 
 
 };
+
+
+
+export const updateCheckOut = async (req, res) => {
+    const { id } = req.params;
+    try {
+        // Find the booking by ID
+        const booking = await bookingModel.findById(id);
+        if (!booking) {
+            return res.status(404).json({ success: false, message: "Booking not found" });
+        }
+
+        // Find the room associated with the booking
+        const room = await roomModel.findById(booking.roomNo);
+        if (!room) {
+            return res.status(404).json({ success: false, message: "Room not found" });
+        }
+
+        // Update room status to "open"
+        room.status = "open";
+        await room.save();
+
+        booking.departDate = new Date();
+        await booking.save();
+
+        return res.status(200).json({
+            success: true,
+            message: "Checkout successful, room status updated to open",
+            booking,
+            room
+        });
+    } catch (error) {
+        return res.status(500).json({ success: false, message: "Something went wrong", error });
+    }
+};
+
